Add junit

+++ personTest/src/main/java/com/pt/person/controller/PersonController.java
@@ -1,9 +1,15 @@
 package com.pt.person.controller;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import com.pt.person.model.Person;
 import com.pt.person.service.PersonService;
@@ -15,8 +21,10 @@ public class PersonController {
     PersonService personService;
 
     @GetMapping("/select")
-    private List<Person> getAllPersons() {
-        return personService.findAll();
+    private ResponseEntity<?> getAllPersons() {
+    	var listPersons = personService.findAll();
+        
+        return ResponseEntity.status(HttpStatus.OK).body(listPersons);
     }
 
     @GetMapping("/select/{id}")
@@ -24,14 +32,20 @@ public class PersonController {
         return personService.getPersonById(id);
     }
 
-    @DeleteMapping("/delete/{id}")
-    private void deletePerson(@PathVariable("id") int id) {
-        personService.delete(id);
-    }
+	@DeleteMapping("/delete/{id}")
+	private ResponseEntity<?> deletePerson(@PathVariable("id") int id) {
+		var user = personService.getPersonById(id);
+		if (user == null) {
+			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
+		} else {
+			personService.delete(id);
+			return ResponseEntity.status(HttpStatus.OK).body(null);
+		}
+	}
 
     @PostMapping("/insert")
-    private int savePerson(@RequestBody Person person) {
-        personService.saveOrUpdate(person);
-        return person.getId();
+    private ResponseEntity<?> savePerson(@RequestBody Person person) {
+    	var user = personService.saveOrUpdate(person);
+        return ResponseEntity.status(HttpStatus.OK).body(user.getId());
     }
 }
